requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
boto3==1.21.27
botocore==1.24.27
cssselect2==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-storages==1.12.3
future==0.18.2
html5lib==1.1
jmespath==1.0.0
lxml==4.8.0
Pillow==9.0.1
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
reportlab==3.6.9
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.63.1
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.6
